+++ administrador/templates/editar_datos_paciente_admin.html
@@ -196,7 +196,7 @@
         alert('¡Cambios Guardados Exitosamente!');
         return true; // Permitir que el formulario se envíe
     }
-    
+
     document.getElementById("boton-cancelar").addEventListener("click", function(){
         window.location.href = "{% url 'admin_pacientes' %}";
     });
